+++ packages/dql/src/ClassGeneration/EntityBasedGeneratorTrait.php
@@ -11,7 +11,7 @@ use Doctrine\ORM\Mapping\ManyToMany;
 use Doctrine\ORM\Mapping\ManyToOne;
 use Doctrine\ORM\Mapping\OneToMany;
 use Doctrine\ORM\Mapping\OneToOne;
-use EDT\Parsing\Utilities\ClassOrInterfaceType;
+use EDT\Parsing\Utilities\Types\ClassOrInterfaceType;
 use InvalidArgumentException;
 use ReflectionAttribute;
 use ReflectionProperty;