+++ packages/jsonapi/tests/Utilities/SingleTemplateClassForProperty.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Utilities;
+
+/**
+ * @template T
+ */
+class SingleTemplateClassForProperty
+{
+
+}