+++ packages/paths/src/PathBuilding/PropertyAutoPathTrait.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace EDT\PathBuilding;
 
 use ArrayIterator;
-use EDT\Parsing\Utilities\TypeInterface;
 use EDT\Parsing\Utilities\ParseException;
-use EDT\Parsing\Utilities\ClassOrInterfaceType;
+use EDT\Parsing\Utilities\Types\ClassOrInterfaceType;
+use EDT\Parsing\Utilities\Types\TypeInterface;
 use EDT\PathBuilding\SegmentFactories\ReflectionSegmentFactory;
 use EDT\PathBuilding\SegmentFactories\SegmentFactoryInterface;
 use EDT\Querying\Contracts\PathException;